chore(tests): introduce vite builder
