+++ tests/playwright/src/index.ts
@@ -1,9 +1,27 @@
+/**********************************************************************
+ * Copyright (C) 2023-2024 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ***********************************************************************/
+
 // export core modules
 export * from './globalSetup/global-setup';
 export * from './runner/podman-desktop-runner';
-export type { RunnerTestContext } from './testContext/runner-test-context';
-export * from './setupFiles/extended-hooks';
+export * from './setupFiles/extended-hooks-utils';
 export * from './setupFiles/setup-registry';
+export type { RunnerTestContext } from './testContext/runner-test-context';
 export * from './utility/cleanup';
 export * from './utility/operations';
 export * from './utility/wait';
