+++ tests/playwright/tsconfig.json
@@ -2,10 +2,11 @@
   "compilerOptions": {
     "module": "esnext",
     "target": "esnext",
-    "sourceMap": false,
+    "sourceMap": true,
     "moduleResolution": "Node",
     "declaration": true,
     "outDir": "dist",
+    "rootDir": "src",
     "skipLibCheck": true,
     "allowSyntheticDefaultImports": true,
     "resolveJsonModule": true,
@@ -20,5 +21,5 @@
     }
   },
   "include": ["src/**/*.ts", "../types/**/*.d.ts"],
-  "exclude": ["node_modules", "src/*spec.ts"]
+  "exclude": ["node_modules", "src/specs/*", "src/setupFiles/extended-hooks.ts"]
 }
